+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/clients/GitHubClient.java
@@ -1,6 +1,6 @@
 package ru.tinkoff.edu.java.scrapper.clients;
 
-import ru.tinkoff.edu.java.scrapper.web.dto.GitHubResponse;
+import ru.tinkoff.edu.java.scrapper.dto.GitHubResponse;
 
 public interface GitHubClient {
     GitHubResponse fetchRepositoryInfo(String owner, String repos);
