+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/clients/StackOverflowClient.java
@@ -1,6 +1,6 @@
 package ru.tinkoff.edu.java.scrapper.clients;
 
-import ru.tinkoff.edu.java.scrapper.web.dto.StackOverflowResponse;
+import ru.tinkoff.edu.java.scrapper.dto.StackOverflowResponse;
 
 public interface StackOverflowClient {
     StackOverflowResponse fetchQuestionInfo(String id);
