+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/controllers/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.scrapper.web.controllers;
+package ru.tinkoff.edu.java.scrapper.controllers;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
-import ru.tinkoff.edu.java.scrapper.web.dto.ApiErrorResponse;
-import ru.tinkoff.edu.java.scrapper.web.exceptions.InvalidRepositoryInformation;
-import ru.tinkoff.edu.java.scrapper.web.exceptions.InvalidQuestionInformation;
+import ru.tinkoff.edu.java.scrapper.dto.ApiErrorResponse;
+import ru.tinkoff.edu.java.scrapper.exceptions.InvalidRepositoryInformation;
+import ru.tinkoff.edu.java.scrapper.exceptions.InvalidQuestionInformation;
 
 @RestControllerAdvice(
     basePackages = "ru.tinkoff.edu.java.scrapper.web.controllers"
@@ -23,7 +23,7 @@ public class GlobalExceptionHandler {
     @ExceptionHandler({MethodArgumentTypeMismatchException.class, MethodArgumentNotValidException.class})
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ResponseEntity<ApiErrorResponse> handleEMethodArgumentsException(Exception e) {
-        return ResponseEntity.status(400)
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST.value())
             .body(
                 ApiErrorResponse.builder()
                     .description("Некорректные параметры запроса")
@@ -34,7 +34,7 @@ public class GlobalExceptionHandler {
                             .map(StackTraceElement::getClassName)
                             .collect(Collectors.toList())
                     )
-                    .code("400")
+                    .code(HttpStatus.BAD_REQUEST.toString())
                     .build()
             );
     }
@@ -42,14 +42,14 @@ public class GlobalExceptionHandler {
     @ExceptionHandler({InvalidRepositoryInformation.class})
     @ResponseStatus(HttpStatus.NOT_FOUND)
     public ResponseEntity<ApiErrorResponse> handleGitHubInvalidInfo(InvalidRepositoryInformation e) {
-        return ResponseEntity.status(404)
+        return ResponseEntity.status(HttpStatus.NOT_FOUND.value())
             .body(
                 ApiErrorResponse.builder()
                     .description("Проверьте еще раз, правильно ли указаны имя владельца и название репозитория")
                     .exceptionName(e.getClass().getSimpleName())
                     .exceptionMessage(e.getMessage())
                     .stacktrace(null)
-                    .code("404")
+                    .code(HttpStatus.NOT_FOUND.toString())
                     .build()
             );
     }
@@ -57,14 +57,14 @@ public class GlobalExceptionHandler {
     @ExceptionHandler({InvalidQuestionInformation.class})
     @ResponseStatus(HttpStatus.NOT_FOUND)
     public ResponseEntity<ApiErrorResponse> handleStackoverflowInfo(InvalidQuestionInformation e) {
-        return ResponseEntity.status(404)
+        return ResponseEntity.status(HttpStatus.NOT_FOUND.value())
             .body(
                 ApiErrorResponse.builder()
                     .description("Некорректно указан id вопроса, он должен быть в цифровом формате")
                     .exceptionName(e.getClass().getSimpleName())
                     .exceptionMessage(e.getMessage())
                     .stacktrace(null)
-                    .code("404")
+                    .code(HttpStatus.NOT_FOUND.toString())
                     .build()
             );
     }
