+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/dto/RemoveLinkRequest.java
@@ -1,12 +1,12 @@
-package ru.tinkoff.edu.java.scrapper.web.dto;
+package ru.tinkoff.edu.java.scrapper.dto;
 
 import org.hibernate.validator.constraints.URL;
 
 import jakarta.validation.constraints.NotNull;
 
-public record AddLinkRequest(
+public record RemoveLinkRequest(
     @NotNull
     @URL
     String link
-) {
-}
+)
+{}
