+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/exceptions/InvalidRepositoryInformation.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.scrapper.web.exceptions;
+package ru.tinkoff.edu.java.scrapper.exceptions;
 
 public class InvalidRepositoryInformation extends Exception {
