--- scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/scheduler/Scheduler.java
@@ -1,6 +1,0 @@
-package ru.tinkoff.edu.java.scrapper.scheduler;
-
-import java.time.Duration;
-
-public record Scheduler(Duration interval) {
-}
